wip: add array interface for ascii, base64.
The data array interface defines behaviour to append floats and ints to the
data array structure. This uses the arrayer interface to generate the
corresponding array, which allows to generate either ascii or base64
arrays. Also, the compressor interface is there, just needs the
implementation for it.

The other changes are caused by introducing this interface, as they now
can use the functions of the provided interfaces directly.