wip on separating functionality of the compressor
Attempt to separate the functionality of the encoder and the compressor
by interoducing differnet implementations to do this. This might
collapse the different array creators (ascii, binary, raw) in the same
structures, by means of simply providing different versions of the
encoder and the compressor.

Functions to add these settings were also added.