Slightly simplified the compressor.
Small refactor to simplify some functions, also renamed some variables.
The current way of writing, we write only single compressed payloads at
a time, thus, we do not need to track an array of sizes for each
payload.

The routines should be modified, such that the compressor only comsumes
a slice of bytes, rather than the full payload.