Refactor array, with ascii, base64, binary files.
Separated the functionality of the encoder, compressor, and appender
interfaces. These now work on the payloads and are not embedded in the
payload itself. For the ascii, base64, and binary formats the
implementations are presented.

The header now controls creating the new arrays, this saves passing
around all the settings of the header file itself. Additionally, the
header is able to contain a separate AppendedData array for the appended
formats. This array is filled on consequentive calls to store data.
Similarly, the DArray has an offset field now, which points into the
AppendedData array. This is all controlled through the payload.

Some options were added to test this behaviour. However, their API needs
to be simplified and/or refactored.