Unified floats and ints in encoder interface.
The encoder interface drops the Ints and Floats implementations to
convert []int or []float into a payload. Now, the encoder interface
requires a binarise routine that accepts an empty interface. Only for
ascii specific implementations are required, as we need to insert the
separating spaces there. For both base64 and raw output, the Ints and
Floats implementations are omitted, as these can be performed by the
newly added NewPayloadFromData routine.

Although it requires an empty interface, it reduces the areas of
repeated code. Additionally, we only really have to deal with type
switches for the ascii encoder. The others, are now directly handles by
binary.Write(...).

Decoding routines need to be added to the interface and the
implementations. Also, add more tests.