The DataArray is now just a struct.
The DataArray is changed into a single struct. There seems to be too
little variation between both implementations to separate this with
interfaces. Now, we just store this struct directly. The switching for
inline and appended storage is done by an if/else.

The raw offset is now a *int, to remove the awkward string and int
conversions we had before. Should only be printed whenever the *int is
explicitly set, as the nil value is ignored by the xml encoding.

Added docs on most structs to clarify their contents.