Allow Data to take interface{}.
To allow the user to pass any type of data, these routines require to
handle the empty interface. This is more general than just allowing
[]float64, as cases might occur where passing ints are desired as well.

Using reflect, the length of these interface{}s are obtained to
determine if the data should be added on points or on cells.

Additionally, the Add() function returns error directly when
encountered. It is up to the user to perform error checking when writing
data. For now, if err != nil, something did go wrong during writing and
requires further inspection.