Extents are represented with bounds.
Added internal bounds type, just being [6]int to represent the lower and
upper bounds for each dimension. This also allows to perform some
validation on the bounds on constructing this type (needs test). Also,
adding the marshaller allows to prevent storing the bounds as both int
and string.

Also, added in error returns for the partitions.