wip regarding coordinates for structuredgrid.
The structured grid requires to provide the coordinates of all points
indivually. Thus, for each point a tuple of x, y, z, is required. The
routine should allow different forms of passing these coordinates. First
of all, one can provide a single slice with all coordinates (xyz)
already appended. Or, one can provide a vector per dimension (x, y, z),
which are then automatically interleaved to match the desired order.
Finally, for two dimensional data, one should be able to supply two out
of three dimensions and the routine is able to fill in the missing zeros
of that missing dimension automatically.

This addds a wip verion of this behaviour, needs verified with testing.