All formats set their points through Points().
In Points we switch depending on the actual format of the header. This
allows that we only need to consider to call Points for each format that
requires storage of the coordinates. For each format that requires
points, i.e. rectilinear, structured, and unstructured, an
implementation is provided to process the received coordinates and write
them to the data structure.