Interleave does not take bounds.
For unstructured grids there are no bounds present, while we still want
to use the interleave and/or splice routines to combine possible
two-dimensional, or separated three-dimensional data, into a single data
vector. The function now takes to int arguments to indicate the expected
number of points and the insertion point. Thus, the function can be used
for unstructured.

Additionally, verifies the length of the components provided to the
interleave and returns an error in case different lengths were observed.